final-demo/requirements.txt
streamlit==1.12.0
numpy
tensorflow-cpu
plotly==5.5.0
streamlit_drawable_canvas